+++ paper.bib
@@ -90,6 +90,7 @@
 	institution = {US Dept. of the Interior, Geological Survey, Water Resources Division,},
 	author = {Colby, BR},
 	year = {1956},
+	doi = {10.3133/ofr5627},
 }
 
 @article{addor_camels_2017,
@@ -117,6 +118,7 @@
 @article{fowler_camels-aus_2021,
 	title = {{CAMELS}-{AUS}: hydrometeorological time series and landscape attributes for 222 catchments in {Australia}},
 	volume = {13},
+	doi = {10.5194/essd-13-3847-2021},
 	number = {8},
 	journal = {Earth System Science Data},
 	author = {Fowler, Keirnan JA and Acharya, Suwash Chandra and Addor, Nans and Chou, Chihchung and Peel, Murray C},
@@ -139,6 +141,7 @@
 @article{arsenault_comprehensive_2020,
 	title = {A comprehensive, multisource database for hydrometeorological modeling of 14,425 {North} {American} watersheds},
 	volume = {7},
+	doi = {10.1038/s41597-020-00583-2},
 	number = {1},
 	journal = {Scientific Data},
 	author = {Arsenault, Richard and Brissette, François and Martel, Jean-Luc and Troin, Magali and Lévesque, Guillaume and Davidson-Chaput, Jonathan and Gonzalez, Mariana Castañeda and Ameli, Ali and Poulin, Annie},
@@ -150,6 +153,7 @@
 @article{linke_global_2019,
 	title = {Global hydro-environmental sub-basin and river reach characteristics at high spatial resolution},
 	volume = {6},
+	doi = {10.1038/s41597-019-0300-6},
 	number = {1},
 	journal = {Scientific data},
 	author = {Linke, Simon and Lehner, Bernhard and Ouellet Dallaire, Camille and Ariwi, Joseph and Grill, Günther and Anand, Mira and Beames, Penny and Burchard-Levine, Vicente and Maxwell, Sally and Moidu, Hana and {others}},
@@ -161,6 +165,7 @@
 @article{kratzert_toward_2019,
 	title = {Toward improved predictions in ungauged basins: {Exploiting} the power of machine learning},
 	volume = {55},
+	doi = {10.1029/2019WR026065},
 	number = {12},
 	journal = {Water Resources Research},
 	author = {Kratzert, Frederik and Klotz, Daniel and Herrnegger, Mathew and Sampson, Alden K and Hochreiter, Sepp and Nearing, Grey S},
@@ -192,6 +197,7 @@
 @article{kratzert_note_2021,
 	title = {A note on leveraging synergy in multiple meteorological data sets with deep learning for rainfall–runoff modeling},
 	volume = {25},
+	doi = {10.5194/hess-25-2685-2021},
 	number = {5},
 	journal = {Hydrology and Earth System Sciences},
 	author = {Kratzert, Frederik and Klotz, Daniel and Hochreiter, Sepp and Nearing, Grey S},
@@ -254,6 +260,7 @@
 @article{gorelick_google_2017-1,
 	title = {Google {Earth} {Engine}: {Planetary}-scale geospatial analysis for everyone},
 	volume = {202},
+	doi = {10.1016/j.rse.2017.06.031},
 	journal = {Remote sensing of Environment},
 	author = {Gorelick, Noel and Hancher, Matt and Dixon, Mike and Ilyushchenko, Simon and Thau, David and Moore, Rebecca},
 	year = {2017},
@@ -273,6 +280,7 @@
 @article{wu_geemap_2020,
 	title = {geemap: {A} {Python} package for interactive mapping with {Google} {Earth} {Engine}},
 	volume = {5},
+	doi = {10.21105/joss.02305},
[…]
