+++ paper.bib
@@ -78,9 +78,11 @@
 	pages = {213--236},
 }
 
-@article{fenton_calculation_2001,
+@techreport{fenton_calculation_2001,
 	title = {The calculation of streamflow from measurements of stage},
+	institution = {Cooperative Research Centre for Catchment Hydrology},
 	author = {Fenton, John D and Keller, Robert John},
+	month = jan,
 	year = {2001},
 	note = {Publisher: CRC for Catchment Hydrology Boca Raton, FL, USA},
 }
@@ -96,6 +98,7 @@
 @article{addor_camels_2017,
 	title = {The {CAMELS} data set: catchment attributes and meteorology for large-sample studies},
 	volume = {21},
+	doi = {10.5194/hess-21-5293-2017},
 	number = {10},
 	journal = {Hydrology and Earth System Sciences},
 	author = {Addor, Nans and Newman, Andrew J and Mizukami, Naoki and Clark, Martyn P},
@@ -107,6 +110,7 @@
 @article{chagas_camels-br_2020,
 	title = {{CAMELS}-{BR}: hydrometeorological time series and landscape attributes for 897 catchments in {Brazil}},
 	volume = {12},
+	doi = {10.5194/essd-12-2075-2020},
 	number = {3},
 	journal = {Earth System Science Data},
 	author = {Chagas, Vinícius BP and Chaffe, Pedro LB and Addor, Nans and Fan, Fernando M and Fleischmann, Ayan S and Paiva, Rodrigo CD and Siqueira, Vinícius A},
@@ -130,6 +134,7 @@
 @article{alvarez-garreton_camels-cl_2018,
 	title = {The {CAMELS}-{CL} dataset: catchment attributes and meteorology for large sample studies–{Chile} dataset},
 	volume = {22},
+	doi = {10.5194/hess-22-5817-2018},
 	number = {11},
 	journal = {Hydrology and Earth System Sciences},
 	author = {Alvarez-Garreton, Camila and Mendoza, Pablo A and Boisier, Juan Pablo and Addor, Nans and Galleguillos, Mauricio and Zambrano-Bigiarini, Mauricio and Lara, Antonio and Puelma, Cristóbal and Cortes, Gonzalo and Garreaud, Rene and {others}},
@@ -177,6 +182,7 @@
 @article{gauch_rainfallrunoff_2021,
 	title = {Rainfall–runoff prediction at multiple timescales with a single {Long} {Short}-{Term} {Memory} network},
 	volume = {25},
+	doi = {10.5194/hess-25-2045-2021},
 	number = {4},
 	journal = {Hydrology and Earth System Sciences},
 	author = {Gauch, Martin and Kratzert, Frederik and Klotz, Daniel and Nearing, Grey and Lin, Jimmy and Hochreiter, Sepp},
@@ -187,6 +193,7 @@
 
 @article{nearing_data_2021,
 	title = {Data assimilation and autoregression for using near-real-time streamflow observations in long short-term memory networks},
+	doi = {10.5194/hess-2021-515},
 	journal = {Hydrology and Earth System Sciences Discussions},
 	author = {Nearing, Grey S and Klotz, Daniel and Sampson, Alden Keefe and Kratzert, Frederik and Gauch, Martin and Frame, Jonathan M and Shalev, Guy and Nevo, Sella},
 	year = {2021},
@@ -320,9 +327,9 @@
 	pages = {5053--5073},
 }
 
-@misc{lehner_hydrobasins_2014,
+@techreport{lehner_hydrobasins_2014,
+	type = {Technical {Report}},
[…]
