### 2022-03-06 #4219 @jonschwenk
Added all the DOIs I could find, including all those specified as missing by editorialbot.
