### 2022-03-08 #4237 @editorialbot
[REVIEW]: rabpro: global watershed boundaries, river elevation profiles, and catchment statistics
**Submitting author:** @jonschwenk (<a href="http://orcid.org/0000-0001-5803-9686">Jon Schwenk</a>)
**Repository:** https://github.com/VeinsOfTheEarth/rabpro
**Branch with paper.md** (empty if default branch):
**Version:** v0.5.2
