### 2022-03-11 VeinsOfTheEarth/rabpro#120 @thurber
JOSS Review
Hello `rabpro` team! I'm one of your JOSS reviewers over in https://github.com/openjournals/joss-reviews/issues/4237. I'll use this issue to keep track of all my feedback, but feel free to create separate tickets if that works best with your workflow!

- [x] the PyPI package version is behind (at 0.2.2), and there may be an issue with a circular dependency (you mention pip install is available in the first paragraph at https://veinsoftheearth.github.io/rabpro/install/index.html); conda-forge installation works!
- [x] at https://veinsoftheearth.github.io/rabpro/configure/index.html, it would be nice to mention what the default download directory will be if a user doesn't specify it with an environment variable -- that way i know if i want to change it or not! (looks like the `data_utils.download_hydrobasins()` method also allows one to specify the path, could mention that too). Update: on Mac, the default path doesn't exist and the download utility throws an error: `FileNotFoundError: [Errno 2] No such file or directory: '/Users/redacted/Library/Application Support/rabpro/HydroBasins.zip'` (similar story with `data_utils.download_merit()`
- [x] downloading with `data_utils.download_hydrobasins('./')` fails with the following message: `Downloading HydroBasins zip file (562 MB)... HydroBasins zip file was not successfully downloaded. Check proxy?` I'm not currently on a network that requires a proxy or VPN so something else must be going wrong -- let me know if you need me to provide more info about this! Update: it looks like the link you share from https://veinsoftheearth.github.io/rabpro/examples/notebooks/downloading_data.html to manually download this data is also access restricted, which may be why the method fails.
- [x] the link at https://veinsoftheearth.github.io/rabpro/configure/index.html to the merit downloads points to the Hydro data but then shows sample code for downloading the DEM data -- turns out those datasets have different user names and passwords 😅

Looks like that's as far as I can get at the moment -- let me know about the HydroBasins data (can I download from hydrosheds directly, or does the data at your google drive link have certain shape/structure that's required?), and I will pick up from here next time!
