### 2022-03-14 VeinsOfTheEarth/rabpro#120 @thurber
- [ ] following along with the basic example at https://veinsoftheearth.github.io/rabpro/examples/notebooks/basic_example.html, I get the following error:
```
>>> rpo.elev_profile(dist_to_walk_km=5)
ERROR 4: /Users/thur961/im3/rabpro/MERIT_Hydro/MERIT_ELEV_HP/MERIT_ELEV_HP.vrt: No such file or directory
ERROR 4: /Users/thur961/im3/rabpro/MERIT_Hydro/MERIT_UDA/MERIT_UDA.vrt: No such file or directory
ERROR 4: /Users/thur961/im3/rabpro/MERIT_Hydro/MERIT_FDR/MERIT_FDR.vrt: No such file or directory
ERROR 4: /Users/thur961/im3/rabpro/MERIT_Hydro/MERIT_WTH/MERIT_WTH.vrt: No such file or directory
Extracting flowpath from DEM...
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
  File "/opt/miniconda3/envs/rabpro/lib/python3.9/site-packages/rabpro/core.py", line 319, in elev_profile
    self.gdf, self.flowline = ep.main(
  File "/opt/miniconda3/envs/rabpro/lib/python3.9/site-packages/rabpro/elev_profile.py", line 89, in main
    cr_ds_mapped, why = mu.map_cl_pt_to_flowline(
  File "/opt/miniconda3/envs/rabpro/lib/python3.9/site-packages/rabpro/merit_utils.py", line 688, in map_cl_pt_to_flowline
    cr = ru.lonlat_to_xy(lonlat[0], lonlat[1], da_obj.GetGeoTransform())
AttributeError: 'NoneType' object has no attribute 'GetGeoTransform'
```

When I look in those directories, i see TIF files but no vrt files:
```
(rabpro) ➜  rabpro ls MERIT_Hydro/MERIT_ELEV_HP
n30w125_elv.tif n50w125_elv.tif n55w135_elv.tif s05e155_elv.tif s10e155_elv.tif s15e155_elv.tif s20e150_elv.tif s20e175_elv.tif s25e170_elv.tif
n35w125_elv.tif n50w130_elv.tif n55w140_elv.tif s05e165_elv.tif s10e160_elv.tif s15e160_elv.tif s20e155_elv.tif s25e150_elv.tif s30e150_elv.tif
n40w125_elv.tif n50w135_elv.tif n55w145_elv.tif s05e170_elv.tif s10e165_elv.tif s15e165_elv.tif s20e160_elv.tif s25e155_elv.tif s30e165_elv.tif
n45w125_elv.tif n55w125_elv.tif n55w150_elv.tif s05e175_elv.tif s10e175_elv.tif s15e170_elv.tif s20e165_elv.tif s25e160_elv.tif
n45w130_elv.tif n55w130_elv.tif s05e150_elv.tif s10e150_elv.tif s15e150_elv.tif s15e175_elv.tif s20e170_elv.tif s25e165_elv.tif
```
