+++ rabpro/data_utils.py
@@ -48,6 +48,41 @@ hydrobasins_paths = {
 }
 
 
+def _path_generator_util(datapath, configpath):
+    if datapath is None:
+        try:
+            datapath = Path(os.environ["RABPRO_DATA"])
+        except:
+            datapath = Path(appdirs.user_data_dir("rabpro", "rabpro"))
+
+    if configpath is None:
+        try:
+            configpath = Path(os.environ["RABPRO_CONFIG"])
+        except:
+            configpath = Path(appdirs.user_config_dir("rabpro", "rabpro"))
+
+    return datapath, configpath
+
+
+def create_file_structure(datapath=None, configpath=None):
+
+    datapath, configpath = _path_generator_util(datapath, configpath)
+
+    os.makedirs(configpath, exist_ok=True)
+    for key in merit_hydro_paths:
+        os.makedirs(os.path.join(datapath, merit_hydro_paths[key]), exist_ok=True)
+
+    for key in hydrobasins_paths:
+        os.makedirs(os.path.join(datapath, hydrobasins_paths[key]), exist_ok=True)
+
+
+def delete_file_structure(datapath=None, configpath=None):
+    datapath, configpath = _path_generator_util(datapath, configpath)
+
+    shutil.rmtree(configpath, ignore_errors=True)
+    shutil.rmtree(datapath, ignore_errors=True)
+
+
 def create_datapaths(datapath=None, configpath=None, reset_user_metadata=False):
     datapath, configpath = _path_generator_util(datapath, configpath)
 
@@ -80,41 +115,6 @@ def create_datapaths(datapath=None, configpath=None, reset_user_metadata=False):
                 json.dump(r, f, indent=4)
 
     return datapaths
-
-
-def create_file_structure(datapath=None, configpath=None):
-
-    datapath, configpath = _path_generator_util(datapath, configpath)
-
-    os.makedirs(configpath, exist_ok=True)
-    for key in merit_hydro_paths:
-        os.makedirs(os.path.join(datapath, merit_hydro_paths[key]), exist_ok=True)
-
-    for key in hydrobasins_paths:
-        os.makedirs(os.path.join(datapath, hydrobasins_paths[key]), exist_ok=True)
-
-
-def delete_file_structure(datapath=None, configpath=None):
-    datapath, configpath = _path_generator_util(datapath, configpath)
-
-    shutil.rmtree(configpath, ignore_errors=True)
-    shutil.rmtree(datapath, ignore_errors=True)
-
-
-def _path_generator_util(datapath, configpath):
-    if datapath is None:
-        try:
-            datapath = Path(os.environ["RABPRO_DATA"])
-        except:
-            datapath = Path(appdirs.user_data_dir("rabpro", "rabpro"))
-
-    if configpath is None:
-        try:
-            configpath = Path(os.environ["RABPRO_CONFIG"])
-        except:
-            configpath = Path(appdirs.user_config_dir("rabpro", "rabpro"))
-
-    return datapath, configpath
 
 
 def does_merit_exist(datapaths):
@@ -192,7 +192,12 @@ def download_hydrobasins(datapath=None, proxy=None):
 
     if datapath is None:
         datapath, _ = _path_generator_util(None, None)
+    else: # Ensure the user-supplied datapath exists, else create
+        if os.path.isdir(datapath) is False:
+            os.path.mkdir(datapath)
+
     datapath = Path(datapath)
+    
     filename = datapath / 'HydroBasins.zip'
     if os.path.isfile(filename):
         os.remove(filename)
@@ -202,7 +207,8 @@ def download_hydrobasins(datapath=None, proxy=None):
     # Check that filesize matches expected
[…]
