### 2022-03-14 VeinsOfTheEarth/rabpro#120 @jonschwenk
Hi @thurber , let me see if I can work through some of these issues.

1. PyPI should not be used for installing rabpro. The `pip` installation specified in the docs is only for a local install (i.e. you cloned the repo). @jsta should we just remove the PyPI version altogether so it's not accidentally installed?

2. (**A**) The [docs](https://veinsoftheearth.github.io/rabpro/configure/index.html) have been updated to demonstrate how to find the default data path `rabpro` expects.
(**B**) If the user supplied a path that does not exist to the `download_XXX()` functions, you would indeed see that error. I added some lines to handle this case--the path will now be created if it does not exist.

4. within-Python downloading issues can be tricky to diagnose. I have expanded the error message to offer the manual download option, e.g.
`HydroBasins zip file was not successfully downloaded. Check proxy? You may also manually download the HydroBasins file from https://drive.google.com/file/d/1NLJUEWhJ9A4y47rcGYv_jWF1Tx2nLEO9/view?usp=sharing and unzip it to C:\redacted\AppData\Local\rabpro\rabpro`
Hopefully that helps clear things up.

5. This was a documentation error, thank you! This is confusing because there are two products: MERIT DEM and MERIT-Hydro (which includes a hydrologically-processed DEM that `rabpro` uses). It is further confusing because `rabpro` contains a download function `download_merit_dem()`, which has no use to rabpro as it needs the MERIT-Hydro (`download_merit_hydro()`) tiles. To make this clearer, I have removed the `download_merit_dem()` function, and I have updated the documentation.

I will address 5. in a separate comment.
