### 2022-03-15 VeinsOfTheEarth/rabpro#120 @jonschwenk
@thurber With respect to missing VRT files, did you manually download the tiles? The `download_merit_hydro` function should (re)build the VRTs after downloading any new tiles.

Either way, I can put a check in the `profiler` class that the VRTs are available and rebuilds them if not. Will update when that is ready.
