### 2022-03-15 VeinsOfTheEarth/rabpro#120 @jsta
Yes, I think we should remove PyPI
