### 2022-03-15 VeinsOfTheEarth/rabpro#120 @thurber
@jonschwenk, i used the `download_merit_hydro` function -- interestingly, i ran it again today and it worked, so i'm not sure what's going on there -- here's the stdout from the non-working and working times, in case that's helpful. maybe just a fluke.

non-working:
```python
>>> data_utils.download_merit_hydro('n30w150', username='redacted', password='redacted')
Downloading 'http://hydro.iis.u-tokyo.ac.jp/~yamadai/MERIT_Hydro/distribute/v1.0/dir_n30w150.tar' into '/Users/redacted/rabpro/MERIT_Hydro/MERIT_FDR/dir_n30w150.tar'
100%|██████████████████████████████████████████████████████████████████████████████████████████████████████████████████████| 72.0M/72.0M [00:08<00:00, 8.07MB/s]
Downloading 'http://hydro.iis.u-tokyo.ac.jp/~yamadai/MERIT_Hydro/distribute/v1.0/elv_n30w150.tar' into '/Users/redacted/rabpro/MERIT_Hydro/MERIT_ELEV_HP/elv_n30w150.tar'
100%|████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████| 866M/866M [01:31<00:00, 9.50MB/s]
Downloading 'http://hydro.iis.u-tokyo.ac.jp/~yamadai/MERIT_Hydro/distribute/v1.0/upa_n30w150.tar' into '/Users/redacted/rabpro/MERIT_Hydro/MERIT_UDA/upa_n30w150.tar'
100%|████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████| 502M/502M [01:20<00:00, 6.19MB/s]
Downloading 'http://hydro.iis.u-tokyo.ac.jp/~yamadai/MERIT_Hydro/distribute/v1.0/wth_n30w150.tar' into '/Users/redacted/rabpro/MERIT_Hydro/MERIT_WTH/wth_n30w150.tar'
100%|██████████████████████████████████████████████████████████████████████████████████████████████████████████████████████| 25.3M/25.3M [00:03<00:00, 6.89MB/s]
>>> rpo.elev_profile(dist_to_walk_km=5)
ERROR 4: /Users/redacted/rabpro/MERIT_Hydro/MERIT_ELEV_HP/MERIT_ELEV_HP.vrt: No such file or directory
ERROR 4: /Users/redacted/rabpro/MERIT_Hydro/MERIT_UDA/MERIT_UDA.vrt: No such file or directory
ERROR 4: /Users/redacted/rabpro/MERIT_Hydro/MERIT_FDR/MERIT_FDR.vrt: No such file or directory
ERROR 4: /Users/redacted/rabpro/MERIT_Hydro/MERIT_WTH/MERIT_WTH.vrt: No such file or directory
Extracting flowpath from DEM...
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
  File "/opt/miniconda3/envs/rabpro/lib/python3.9/site-packages/rabpro/core.py", line 319, in elev_profile
    self.gdf, self.flowline = ep.main(
  File "/opt/miniconda3/envs/rabpro/lib/python3.9/site-packages/rabpro/elev_profile.py", line 89, in main
    cr_ds_mapped, why = mu.map_cl_pt_to_flowline(
  File "/opt/miniconda3/envs/rabpro/lib/python3.9/site-packages/rabpro/merit_utils.py", line 688, in map_cl_pt_to_flowline
    cr = ru.lonlat_to_xy(lonlat[0], lonlat[1], da_obj.GetGeoTransform())
AttributeError: 'NoneType' object has no attribute 'GetGeoTransform'
>>>
```

working:
```python
>>> data_utils.download_merit_hydro('n30w150', username='hydrography', password='rivernetwork')
Downloading 'http://hydro.iis.u-tokyo.ac.jp/~yamadai/MERIT_Hydro/distribute/v1.0/dir_n30w150.tar' into '/Users/redacted/rabpro/MERIT_Hydro/MERIT_FDR/dir_n30w150.tar'
100%|██████████████████████████████████████████████████████████████████████████████████████████████████████████████████████| 72.0M/72.0M [00:10<00:00, 6.95MB/s]
Downloading 'http://hydro.iis.u-tokyo.ac.jp/~yamadai/MERIT_Hydro/distribute/v1.0/elv_n30w150.tar' into '/Users/redacted/rabpro/MERIT_Hydro/MERIT_ELEV_HP/elv_n30w150.tar'
100%|████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████| 866M/866M [01:37<00:00, 8.92MB/s]
Downloading 'http://hydro.iis.u-tokyo.ac.jp/~yamadai/MERIT_Hydro/distribute/v1.0/upa_n30w150.tar' into '/Users/redacted/rabpro/MERIT_Hydro/MERIT_UDA/upa_n30w150.tar'
100%|████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████| 502M/502M [00:55<00:00, 9.00MB/s]
Downloading 'http://hydro.iis.u-tokyo.ac.jp/~yamadai/MERIT_Hydro/distribute/v1.0/wth_n30w150.tar' into '/Users/redacted/rabpro/MERIT_Hydro/MERIT_WTH/wth_n30w150.tar'
100%|██████████████████████████████████████████████████████████████████████████████████████████████████████████████████████| 25.3M/25.3M [00:04<00:00, 5.83MB/s]
Building flow direction virtual raster DEM from MERIT tiles...
0...10...20...30...40...50...60...70...80...90...100 - done.

Building drainage areas virtual raster DEM from MERIT tiles...
0...10...20...30...40...50...60...70...80...90...100 - done.

Building hydrologically-processed elevations virtual raster DEM from MERIT tiles...
0...10...20...30...40...50...60...70...80...90...100 - done.

Building width virtual raster from MERIT tiles...
0...10...20...30...40...50...60...70...80...90...100 - done.
>>> rpo.elev_profile(dist_to_walk_km=5)
Extracting flowpath from DEM...
>>>
```
