### 2022-03-15 VeinsOfTheEarth/rabpro#120 @thurber
- [ ] at this page https://veinsoftheearth.github.io/rabpro/examples/notebooks/multiple_basins.html, `conda install contextily` should be `conda install -c conda-forge contextily`
- [ ] in the multiple basins example (https://veinsoftheearth.github.io/rabpro/examples/notebooks/multiple_basins.html), you mention being able to skip the GCP bucket step since you provide a public asset; i think maybe the steps 5 onward assume i did not skip that step -- is it possible to test without creating the GCP assets? (my free trial is over 😿)
