### 2022-03-15 84600cf
addressing #120 virtual rasters
