+++ rabpro/basin_stats.py
@@ -496,7 +496,7 @@ def _get_controls(datasets):
     """
 
     # Load raster metadata file
-    datapaths = ru.get_datapaths(rebuild_vrts=False)
+    datapaths = ru.get_datapaths()
     with open(datapaths["gee_metadata"]) as json_file:
         datadict = {d["id"]: d for d in json.load(json_file)}
 
