+++ tests/test_rebuild_vrts.py
@@ -8,19 +8,22 @@ def test_rebuild_vrts():
     # initial building
     f = io.StringIO()
     with contextlib.redirect_stdout(f):
-        dps = ru.get_datapaths(rebuild_vrts=True, quiet=False)
+        dps = ru.get_datapaths()
+        ru.build_virtual_rasters(dps, skip_if_exists=False, verbose=True)
     output = f.getvalue()
 
     # ensure caching has occurred
     f = io.StringIO()
     with contextlib.redirect_stdout(f):
-        dps = ru.get_datapaths(rebuild_vrts=True, quiet=False)
+        dps = ru.get_datapaths()
+        ru.build_virtual_rasters(dps, skip_if_exists=False, verbose=False)
     output = f.getvalue()
     assert len(output) == 0
 
     # ensure cache override works
     f = io.StringIO()
     with contextlib.redirect_stdout(f):
-        dps = ru.get_datapaths(rebuild_vrts=True, quiet=False, force=True)
+        dps = ru.get_datapaths(force=True)
+        ru.build_virtual_rasters(dps, skip_if_exists=False, verbose=True)
     output = f.getvalue()
     assert len(output) > 0
