### 2022-03-15 VeinsOfTheEarth/rabpro#120 @jonschwenk
Re: contextily, I have updated that notebook/doc to include `-c conda-forge`

Re: skipping the GCP bucket step--no, unfortunately you can't test it without a GCP account. However, that notebook states that `This can be done manually through the **GEE code editor** or via rabpro automation.` In fact, most users would likely want to use the GEE code editor as it is much simpler/user-friendly. The rabpro automation is really designed for more complex raster uploads (`imageCollections` rather than `images`, images with multiple bands, etc.), and is provided as a convenience more than integral functionality. That said, I can send you credentials if you'd like to use our bucket--it's free for another month :)
