+++ _sources/examples/notebooks/multiple_basins.ipynb.txt
@@ -179,9 +179,9 @@
    "id": "ee472c6a-77a1-405b-8357-6f363b386176",
    "metadata": {},
    "source": [
-    "We need to upload our basin vector layer as a GEE asset because it too large to send as a json payload. This can be done manually through the GEE code editor or via `rabpro` automation. This automation requires that you have a writeable Google Cloud Platform (GCP) bucket and that you are authenticated via the command-line to call the `gsutil` and `earthengine` programs. These programs enable sending files to GCP and onward to GEE respectively.\n",
+    "We need to upload our basin vector layer as a GEE asset because it too large to send as a json payload. This can be done either (1) manually through the GEE code editor, but you must first export as a shapefile, or (2) via `rabpro` automation. This automation requires that you have a writeable Google Cloud Platform (GCP) bucket and that you are authenticated via the command-line to call the `gsutil` and `earthengine` programs. These programs enable sending files to GCP and onward to GEE respectively.\n",
     "\n",
-    "For now, you can skip this step as we've uploaded the sl_basins file as a public asset."
+    "For now, you can skip this step as we've uploaded the sl_basins file as a public asset. However, for instructions on uploading your shapefile as a GEE asset, see [this page].(https://developers.google.com/earth-engine/cloud/assets#uploading-assets)."
    ]
   },
   {
