### 2022-03-15 VeinsOfTheEarth/rabpro#120 @thurber
My point of confusion was thinking I could skip the step about creating my own GCP bucket -- maybe I can? In either case, I'm getting some errors relating to the interaction with `gsutil`. I have installed and initialized `gsutil`, but might certainly be missing some steps. Any ideas?

Trying to upload to my own bucket:
```python
>>> your_gee_asset = rabpro.utils.upload_gee_vector_asset(
...     zip_path, gee_user, gcp_bucket, gcp_folder="rabpro"
... )
gsutil cp basins_sl.zip tbt-rabpro-tmp/rabpro/basins_sl.zip
Traceback (most recent call last):
  File "/opt/miniconda3/envs/rabpro/lib/python3.9/site-packages/rabpro/utils.py", line 944, in upload_gee_vector_asset
    subprocess.call(shell_cmd)
  File "/opt/miniconda3/envs/rabpro/lib/python3.9/subprocess.py", line 349, in call
    with Popen(*popenargs, **kwargs) as p:
  File "/opt/miniconda3/envs/rabpro/lib/python3.9/subprocess.py", line 951, in __init__
    self._execute_child(args, executable, preexec_fn, close_fds,
  File "/opt/miniconda3/envs/rabpro/lib/python3.9/subprocess.py", line 1821, in _execute_child
    raise child_exception_type(errno_num, err_msg, err_filename)
FileNotFoundError: [Errno 2] No such file or directory: 'gsutil cp basins_sl.zip tbt-rabpro-tmp/rabpro/basins_sl.zip'

Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
  File "/opt/miniconda3/envs/rabpro/lib/python3.9/site-packages/rabpro/utils.py", line 946, in upload_gee_vector_asset
    raise RuntimeError(
RuntimeError: Errors here could indicate that gsutil is not installed.
```

Trying to use the public file:
```python
>>> your_gee_asset = 'users/jstacompute/basins_sl'
>>>
>>>
>>> if not os.path.exists("res.gpkg"):
...     urls, tasks = rabpro.basin_stats.compute(
...         dataset_list, gee_feature_path=your_gee_asset, folder="rabpro"
...     )
...     tag_list = ["wateroccur", "temperature_2m", "precip"]
...     data = rabpro.basin_stats.fetch_gee(urls, tag_list)
...     res = gpd.GeoDataFrame(data.merge(basins_sl, on='vote_id'))
...     res = res.set_geometry("geometry")
...     res.to_file("res.gpkg", driver="GPKG")
... else:
...     res = gpd.read_file("res.gpkg")
...
Traceback (most recent call last):
  File "/opt/miniconda3/envs/rabpro/lib/python3.9/site-packages/ee/data.py", line 330, in _execute_cloud_call
    return call.execute(num_retries=num_retries)
  File "/opt/miniconda3/envs/rabpro/lib/python3.9/site-packages/googleapiclient/_helpers.py", line 134, in positional_wrapper
    return wrapped(*args, **kwargs)
  File "/opt/miniconda3/envs/rabpro/lib/python3.9/site-packages/googleapiclient/http.py", line 915, in execute
    raise HttpError(resp, content, uri=self.uri)
googleapiclient.errors.HttpError: <HttpError 400 when requesting https://earthengine.googleapis.com/v1alpha/projects/earthengine-legacy/table:export?alt=json returned "Collection.loadTable: Collection asset 'users/jstacompute/basins_sl' not found.". Details: "Collection.loadTable: Collection asset 'users/jstacompute/basins_sl' not found.">

Traceback (most recent call last):
  File "<stdin>", line 2, in <module>
  File "/opt/miniconda3/envs/rabpro/lib/python3.9/site-packages/rabpro/basin_stats.py", line 397, in compute
    task.start()
  File "/opt/miniconda3/envs/rabpro/lib/python3.9/site-packages/ee/batch.py", line 89, in start
    result = data.exportTable(self._request_id, self.config)
  File "/opt/miniconda3/envs/rabpro/lib/python3.9/site-packages/ee/data.py", line 1319, in exportTable
    return _prepare_and_run_export(
  File "/opt/miniconda3/envs/rabpro/lib/python3.9/site-packages/ee/data.py", line 1407, in _prepare_and_run_export
    return _execute_cloud_call(
  File "/opt/miniconda3/envs/rabpro/lib/python3.9/site-packages/ee/data.py", line 332, in _execute_cloud_call
    raise _translate_cloud_exception(e)
ee.ee_exception.EEException: Collection.loadTable: Collection asset 'users/jstacompute/basins_sl' not found.
```
