### 2022-03-15 VeinsOfTheEarth/rabpro#120 @jonschwenk
Ah, I see. The user can (and should, probably) skip the bucket upload when working through that example. I will doubly-clarify that in the docs.

Oops, we forgot to make that asset public. Just shared it, so try again and it should work!
