### 2022-03-15 VeinsOfTheEarth/rabpro#120 @thurber
Thanks, that worked!

One more issue in this notebook:

- [ ] The method signatures for the plotting must have changed at some point, i get some errors regarding this line `legend_kwds={"label": col_label, "orientation": "horizontal", "pad": 0.01},`

```python
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
  File "<stdin>", line 2, in <listcomp>
  File "<stdin>", line 5, in build_panel
  File "/opt/miniconda3/envs/rabpro/lib/python3.9/site-packages/geopandas/plotting.py", line 950, in __call__
    return plot_dataframe(data, *args, **kwargs)
  File "/opt/miniconda3/envs/rabpro/lib/python3.9/site-packages/geopandas/plotting.py", line 926, in plot_dataframe
    ax.legend(patches, categories, **legend_kwds)
  File "/opt/miniconda3/envs/rabpro/lib/python3.9/site-packages/matplotlib/axes/_axes.py", line 305, in legend
    self.legend_ = mlegend.Legend(self, handles, labels, **kwargs)
TypeError: __init__() got an unexpected keyword argument 'label'
```

```python
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
  File "<stdin>", line 2, in <listcomp>
  File "<stdin>", line 5, in build_panel
  File "/opt/miniconda3/envs/rabpro/lib/python3.9/site-packages/geopandas/plotting.py", line 950, in __call__
    return plot_dataframe(data, *args, **kwargs)
  File "/opt/miniconda3/envs/rabpro/lib/python3.9/site-packages/geopandas/plotting.py", line 926, in plot_dataframe
    ax.legend(patches, categories, **legend_kwds)
  File "/opt/miniconda3/envs/rabpro/lib/python3.9/site-packages/matplotlib/axes/_axes.py", line 305, in legend
    self.legend_ = mlegend.Legend(self, handles, labels, **kwargs)
TypeError: __init__() got an unexpected keyword argument 'orientation'
```

```python
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
  File "<stdin>", line 2, in <listcomp>
  File "<stdin>", line 5, in build_panel
  File "/opt/miniconda3/envs/rabpro/lib/python3.9/site-packages/geopandas/plotting.py", line 950, in __call__
    return plot_dataframe(data, *args, **kwargs)
  File "/opt/miniconda3/envs/rabpro/lib/python3.9/site-packages/geopandas/plotting.py", line 926, in plot_dataframe
    ax.legend(patches, categories, **legend_kwds)
  File "/opt/miniconda3/envs/rabpro/lib/python3.9/site-packages/matplotlib/axes/_axes.py", line 305, in legend
    self.legend_ = mlegend.Legend(self, handles, labels, **kwargs)
TypeError: __init__() got an unexpected keyword argument 'pad'
```
