### 2022-03-16 VeinsOfTheEarth/rabpro#120 @jonschwenk
@thurber, re:

> the manual download is working for me now, thanks! the data_utils.download_hydrobasins() is still showing the same error though

At the cost of an additional dependency, I replaced the HydroBasins download function (which used the `requests` library) with functionality provided by `pydown`, which seems to be an actively supported package. It makes the code much simpler and more robust. If you update your `rabpro` install to the repo's latest version, it should work now (unless the root problem was deeper). I, too, was getting mixed results with the original download function, but the `pydown`-based one seems to work every time.
