### 2022-03-17 VeinsOfTheEarth/rabpro#120 @thurber
thanks @jonschwenk! With the latest rabpro from main, there may be missing a missing dependency on `gdown` (although this could be from how my environment is now a mix of conda and pip...) -- just double check that.

With these updates, i'm able to get through all the examples, thanks! I'll focus on the paper itself now.
