### 2022-03-17 VeinsOfTheEarth/rabpro#120 @jonschwenk
I updated the [environment file](https://github.com/VeinsOfTheEarth/rabpro/blob/main/environment.yml) to include `gdown`, so we're all good there. Once the _JOSS_ review is complete, we'll repackage for conda-forge to incorporate all the changes for the conda installation.
