+++ rabpro/data_utils.py
@@ -90,10 +90,10 @@ def create_datapaths(datapath=None, configpath=None, reset_user_metadata=False):
     datapath, configpath = _path_generator_util(datapath, configpath)
 
     datapaths = {key: str(datapath / Path(val)) for key, val in _PATH_CONSTANTS.items()}
-    
-    datapaths['HydroBasins_root'] = str(datapath / Path('HydroBasins'))
-    datapaths['MERIT_root'] = str(datapath / Path('MERIT_Hydro'))
-    datapaths['root'] = str(datapath)
+
+    datapaths["HydroBasins_root"] = str(datapath / Path("HydroBasins"))
+    datapaths["MERIT_root"] = str(datapath / Path("MERIT_Hydro"))
+    datapaths["root"] = str(datapath)
 
     gee_metadata_path = datapath / "gee_datasets.json"
     datapaths["gee_metadata"] = str(gee_metadata_path)
@@ -128,15 +128,19 @@ def does_merit_exist(datapaths):
     """
     vrts_exist = 0
     geotiffs_exist = 0
-    dem_files = [k for k in datapaths.keys() if 'DEM' in k]
+    dem_files = [k for k in datapaths.keys() if "DEM" in k]
     for df in dem_files:
-        geotiffs = [f for f in os.listdir(os.path.dirname(datapaths[df])) if f.split('.')[-1] == 'tif']
+        geotiffs = [
+            f
+            for f in os.listdir(os.path.dirname(datapaths[df]))
+            if f.split(".")[-1] == "tif"
+        ]
         if len(geotiffs) > 0:
             geotiffs_exist = geotiffs_exist + 1
         if os.path.isfile(datapaths[df]) is True:
             vrts_exist = vrts_exist + 1
-    
-    return geotiffs_exist, vrts_exist 
+
+    return geotiffs_exist, vrts_exist
 
 
 def does_hydrobasins_exist(datapaths):
@@ -144,11 +148,21 @@ def does_hydrobasins_exist(datapaths):
     Checks if level 1 and level 12 HydroBasins data are available.
     """
     lev1, lev12 = False, False
-    if os.path.isfile(os.path.join(datapaths['HydroBasins1'], 'hybas_all_lev01_v1c.shp')) is True:
+    if (
+        os.path.isfile(
+            os.path.join(datapaths["HydroBasins1"], "hybas_all_lev01_v1c.shp")
+        )
+        is True
+    ):
         lev1 = True
-    if os.path.isfile(os.path.join(datapaths['HydroBasins12'], 'hybas_af_lev12_v1c.shp')) is True:
+    if (
+        os.path.isfile(
+            os.path.join(datapaths["HydroBasins12"], "hybas_af_lev12_v1c.shp")
+        )
+        is True
+    ):
         lev12 = True
-    return lev1, lev12 
+    return lev1, lev12
 
 
 def download_gee_metadata(datapath=None):
@@ -171,16 +185,16 @@ def download_gee_metadata(datapath=None):
             print(
                 f"{CATALOG_URL} download error. Place manually into {gee_metadata_path}"
             )
-            
-            
+
+
 def _download_file_from_google_drive(id_file, destination, proxy=None):
     """
     Self-explanatory.
     """
-    
-    url = 'https://drive.google.com/uc?id={}'.format(id_file)
+
+    url = "https://drive.google.com/uc?id={}".format(id_file)
     gdown.download(url, output=destination, proxy=proxy)
-    
+
     return
 
 
@@ -206,39 +220,43 @@ def download_hydrobasins(datapath=None, proxy=None):
 
     if datapath is None:
         datapath, _ = _path_generator_util(None, None)
-    else: # Ensure the user-supplied datapath exists, else create
[…]
