### 2022-03-22 ec5a761
option for skipping gee metadata download
