### 2022-04-18 #4237 @thurber
hi @crvernon, all my initial software feedback has been addressed by @jonschwenk et al. I still need to take a detailed read of the paper to address the remaining checkboxes; i'll plan to wrap that up this week!
