### 2022-04-25 d91e141
rm rivgraph leftovers, #48
