+++ .github/workflows/build.yaml
@@ -20,7 +20,7 @@ jobs:
       fail-fast: false
       matrix:
         os: [ubuntu-latest , macos-latest] # windows-latest
-        python-version: [3.8] #[3.7, 3.8, 3.9] # undo once rivgraph dep removed
+        python-version: [3.9]
         include:
           # All paths are relative to $HOME
           - os: ubuntu-latest
@@ -60,7 +60,7 @@ jobs:
           miniconda-version: "latest"
           activate-environment: rabpro
           environment-file: environment.yml
-          channels: jschwenk, conda-forge, defaults
+          channels: conda-forge
           auto-activate-base: false
           use-only-tar-bz2: true # needs to be set for caching to work
 
