### 2022-05-06 VeinsOfTheEarth/rabpro#120 @thurber
Hi all, sorry for long delay! Reading through the paper now; great work! Here are some feedback items:
- [ ] In the figure 1 caption, a couple of the references aren't rendering correctly in the compiled paper (version i'm looking at [here](https://github.com/openjournals/joss-papers/blob/joss.04237/joss.04237/10.21105.joss.04237.pdf))
- [ ] In the Statement of Need, the reference to "Brooks (2003)" is in a different style than the rest
- [ ] Any chance there are citations already available for the ongoing rabpro work mentioned towards the end of Statement of Need? No worries if not, but sounds fascinating!
- [ ] Some inconsistencies and broken links for the citations in State of the Field

Nitpicks (up to you whether to change or not):
- [ ] "subbasin-based" is an awkward compound word (~ line 77)
- [ ] "unioning" feels awkward in this sentence (~ line 86).
- [ ] "the cleanest workflow would be" --> "the cleanest workflow is to" (~ line 92)
- [ ] "the workflow would" --> "the workflow is to" (~ line 94)
