+++ paper.bib
@@ -489,7 +489,7 @@
 	keywords = {population},
 }
 
-@misc{didan__kamel_mod13a2_2015,
+@misc{didankamel_mod13a2_2015,
 	title = {{MOD13A2} {MODIS}/{Terra} {Vegetation} {Indices} 16-{Day} {L3} {Global} 1km {SIN} {Grid} {V006}},
 	url = {https://lpdaac.usgs.gov/products/mod13a2v006/},
 	urldate = {2022-02-26},
