+++ paper.bib
@@ -547,6 +547,6 @@ Publisher: Nature Publishing Group},
 @techreport{copernicus_climate_change_service_era5_2017,
 	title = {{ERA5}: {Fifth} generation of {ECMWF} atmospheric reanalyses of the global climate . {Copernicus} {Climate} {Change} {Service} {Climate} {Data} {Store} ({CDS})},
 	url = {https://cds.climate.copernicus.eu/cdsapp#!/home},
-	author = {Copernicus Climate Change Service},
+	author = {Copernicus},
 	year = {2017},
 }
