+++ paper.md
@@ -45,7 +45,7 @@ While we point to streamflow modeling as an example, many other applications exi
 
 ## State of the field
 
-The importance of watersheds, availability of DEMs, and growing computational power has led to the development of many excellent open-source terrain (DEM) analysis packages that provide watershed delineation tools, including [TauDEM](https://hydrology.usu.edu/taudem/taudem5/) [@tarboton_terrain_2005], [pysheds](https://github.com/mdbartos/pysheds) [@bartos_pysheds_2020], [Whitebox Tools][https://github.com/jblindsay/whitebox-tools](@lindsay_whitebox_2016), [SAGA](https://sagatutorials.wordpress.com/terrain-analysis/) [@conrad_system_2015], among many others. Computing statistics and forcing data from geospatial rasters also has a rich history of development, and Google Earth Engine [@gorelick_google_2017] has played an important role. Almost a decade has passed since Google Earth Engine has been available to developers, and the community has in-turn developed open-source packages to interface with its Python API in user-friendlier ways, including [gee_tools](https://github.com/gee-community/gee_tools) [@principe_gee_tools_2021], [geemap](https://geemap.org/) [@wu_geemap_2020], [eemont](https://github.com/davemlz/eemont) [@montero_eemont_2021], and [restee](https://github.com/KMarkert/restee) [@markert_restee_2021]–each of which provides support for sampling zonal statistics and time series from geospatial polygons.
+The importance of watersheds, availability of DEMs, and growing computational power has led to the development of many excellent open-source terrain (DEM) analysis packages that provide watershed delineation tools, including [TauDEM](https://hydrology.usu.edu/taudem/taudem5/) [@tarboton_terrain_2005], [pysheds](https://github.com/mdbartos/pysheds) [@bartos_pysheds_2020], [Whitebox Tools](https://github.com/jblindsay/whitebox-tools) [@lindsay_whitebox_2016], [SAGA](https://sagatutorials.wordpress.com/terrain-analysis/) [@conrad_system_2015], among many others. Computing statistics and forcing data from geospatial rasters also has a rich history of development, and Google Earth Engine [@gorelick_google_2017] has played an important role. Almost a decade has passed since Google Earth Engine has been available to developers, and the community has in-turn developed open-source packages to interface with its Python API in user-friendlier ways, including [gee_tools](https://github.com/gee-community/gee_tools) [@principe_gee_tools_2021], [geemap](https://geemap.org/) [@wu_geemap_2020], [eemont](https://github.com/davemlz/eemont) [@montero_eemont_2021], and [restee](https://github.com/KMarkert/restee) [@markert_restee_2021]–each of which provides support for sampling zonal statistics and time series from geospatial polygons.
 
 However, to our knowledge, `rabpro` is the only available package that provides efficient end-to-end delineation and characterization of watershed basins at scale. While a combination of the cited terrain analysis packages and GEE toolboxes can achieve `rabpro`’s functionality, `rabpro`’s blending of them enables simpler, less error-prone, and faster results.
 
