### 2022-05-20 e2ff8d3
allowing holes in perimeter calcs
