### 2022-05-21 VeinsOfTheEarth/rabpro#125 @kanishkan91
JOSS review : Create zenodo entry for hydro basins file
Currently in the repository, specifically the data instructions- (https://veinsoftheearth.github.io/rabpro/examples/notebooks/downloading_data.html#Download-HydroBasins)

, you are provoding a link to a google drive link to download the Hydrobasins file. This is a bit arbitrary. Please put this on zenodo so there is an archived/minted version of this file. It will also automatically control for versioning.

https://github.com/openjournals/joss-reviews/issues/4237
