### 2022-05-22 #4237 @kanishkan91
@jonschwenk I have started my review. Though I want to raise a very basic question here (just in case I am missing something). From the documentation and the code I gather that you are using the HydroBasins data product to produce aggregate delineations. But is it that rabpro can _only_ be used with HydroBasins for producing more aggregate delineations? For example, if I had my own aggregate basins file, can I use rabpro with it (Maybe by just switching in the shape file)?

The reason I raise this is that there are Earth System Models that use custom basin boundaries. If this package can be used with such custom basin files, it would be extremely useful to the community.
