### 2022-05-22 VeinsOfTheEarth/rabpro#127 @kanishkan91
JOSS review: Make environment.yml consistent with setup.py
The two files are not consistent. For example, `gdown` is missing from setup.py. I created a conda environment in pycharm and then cloned the github repo. It missed the gdown package which I had to install separately.

https://github.com/openjournals/joss-reviews/issues/4237
