### 2022-05-22 VeinsOfTheEarth/rabpro#128 @kanishkan91
JOSS review: Add an illustrative version of MERIT tile along with HydroBasins file? 
In the current documentation a user is asked to create an account for the MERIT files and then the code here (https://veinsoftheearth.github.io/rabpro/configure/index.html#downloading-merit-hydro) provides an example of how you can download the data. For the sake of user convenience , can you also add a sample file to the repo or to the downlood link with the HydroBasins file? That way, the user when just wanting to look at code would always have access to a sample. You can provide a citation for the file provided in the repo. This would be in addition to the extensive instructions already provided (The current instructions are very useful in knowing how to download from the online MERIT  database).

https://github.com/openjournals/joss-reviews/issues/4237
