### 2022-05-22 VeinsOfTheEarth/rabpro#129 @kanishkan91
JOSS review: Request authors to comment on test coverage
I would like the authors to comment on the test coverage. I pulled up the coverage statistics (See image below). I see that the utils files have low coverage by tests (not an issue). But some key files like the basin_stats have low coverage. Could you comment on that? I will look into some sample lines in more detail and maybe this is not an issue, but just wanted to point out.

![image](https://user-images.githubusercontent.com/37234086/169709140-7e1a6988-3f5b-4824-89ef-f60a0034b7c1.png)

https://github.com/openjournals/joss-reviews/issues/4237
