### 2022-05-22 VeinsOfTheEarth/rabpro#130 @kanishkan91
JOSS review: General documentation
- [x] Move the contribution instructions to the main readme in addition to where they are in the documentation
- [x]  Same for installation instructions
- [x]  Do the %capture (see image below) in the documentation signify anything? If not, can those be taken out?

![image](https://user-images.githubusercontent.com/37234086/169856775-43ccfe8a-7fac-4614-8d77-d27e7ab37bfe.png)

- [x] Mention the usage of rabpro.utils.coords_to_merit_tile() on the Basic example page (maybe just add it to the code example there - https://veinsoftheearth.github.io/rabpro/examples/notebooks/basic_example.html). Its an important function IMO.

https://github.com/openjournals/joss-reviews/issues/4237
