### 2022-05-23 VeinsOfTheEarth/rabpro#125 @jonschwenk
Hi @kanishkan91, thanks for reviewing!

HydroBASINs is hosted by its creators at https://www.hydrosheds.org/products/hydrobasins. We provided the Google Drive file only as a convenience--to make the downloading and pathing as simple as possible. I have updated the [docs ](https://veinsoftheearth.github.io/rabpro/configure/index.html#basin-delineation) to describe how to download this data from either the native provider or via the rabpro function.

(As early as a year or so ago, it was hard to find HydroBASINs data--Hydrosheds has since created a nice website that appears to host it more permanently.)
