### 2022-05-23 VeinsOfTheEarth/rabpro#131 @kanishkan91
JOSS review: Is there a map/code somewhere which shows/maps a user coordinate to a corresponding MERIT tile? 
Let me know if I am missing something here. In the workflow example here- https://veinsoftheearth.github.io/rabpro/examples/notebooks/basic_example.html, a user can delinate a watershed based on random co-ordinates. But they would have to read in the corresponding tile from the MERIT files. I know that regular users would have all the MERIT tiles processed. But is there a way, a user can locate the corresponding MERIT tile for a co-ordinate? And in case you have a coordinate that does not have a corresponding MERIT tile for some reason, you get a warning or error? I think this would be a useful addition even for test cases and such.

Please only consider this a suggestion in case I am missing something fundamental.

https://github.com/openjournals/joss-reviews/issues/4237
