### 2022-05-23 VeinsOfTheEarth/rabpro#129 @jsta
@kanishkan91 Where did you pull the coverage statistics from? I'm seeing much higher percentages in the latest CI runs:
![Screenshot 2022-05-23 102227](https://user-images.githubusercontent.com/7844578/169864191-ee6bd32f-518e-4247-95f0-e9ce56d9f979.png)

https://github.com/VeinsOfTheEarth/rabpro/runs/6530778811?check_suite_focus=true#step:8:87
