### 2022-05-23 VeinsOfTheEarth/rabpro#129 @kanishkan91
@jsta That is strange. I just pulled the stats from my local repo using `coverage`. I do see the higher coverage in the CI files. Thanks very much!

Let me look into it again.
