### 2022-05-23 VeinsOfTheEarth/rabpro#131 @kanishkan91
@jonschwenk This all sounds good to me. If possible, could you also mention the usage of `rabpro.utils.coords_to_merit_tile()` on the workflow page (maybe just add it to the code example there)? Its an important function IMO.

Also, sorry to fixate onb this more but does this function- `rabpro.utils.coords_to_merit_tile()` throw a warning when it can't find a tile? Even that would be very useful.
