+++ rabpro/utils.py
@@ -51,7 +51,9 @@ def envvars_rabpro():
     return res_dict
 
 
-def get_datapaths(root_path=None, config_path=None, force=False, update_gee_metadata=False):
+def get_datapaths(
+    root_path=None, config_path=None, force=False, update_gee_metadata=False
+):
     """
     Returns a dictionary of paths to all data that rabpro uses. Also builds
     virtual rasters for MERIT data and downloads latest version of GEE catalog.
@@ -60,12 +62,12 @@ def get_datapaths(root_path=None, config_path=None, force=False, update_gee_meta
     ----------
     root_path: string, optional
         Path to rabpro Data folder that contains the HydroBASINS, MERIT-Hydro,
-        and/or gee catalog jsons. Will read from an environment variable 
-        "RABPRO_DATA". If this variable is not set, uses appdirs to create a 
-        local data directory. This path is the parent directory for the MERIT 
+        and/or gee catalog jsons. Will read from an environment variable
+        "RABPRO_DATA". If this variable is not set, uses appdirs to create a
+        local data directory. This path is the parent directory for the MERIT
         and HydroBasins data directories.
     config_path: string, optional
-        Path to rabpro config folder. Will read from an environment variable 
+        Path to rabpro config folder. Will read from an environment variable
         "RABPRO_CONFIG". If not set, uses appdirs to create local directory.
     force: boolean, optional
         Set True to override datapath caching. Otherwise only fetched once per py session.
@@ -82,7 +84,7 @@ def get_datapaths(root_path=None, config_path=None, force=False, update_gee_meta
     .. code-block:: python
 
         from rabpro import utils
-        utils.get_datapaths()        
+        utils.get_datapaths()
     """
 
     # This chunk makes sure that folder creation, data downloads, etc. only
@@ -103,7 +105,7 @@ def get_datapaths(root_path=None, config_path=None, force=False, update_gee_meta
 def build_virtual_rasters(datapaths, skip_if_exists=False, verbose=True, **kwargs):
     """
     Builds virtual rasters on the four MERIT-Hydro tilesets.
-    
+
     Parameters
     ----------
     datapaths: dict
@@ -397,7 +399,7 @@ def raster_extents(raster_path):
     Returns
     -------
     list
-        [xmin, xmax, ymin, ymax] 
+        [xmin, xmax, ymin, ymax]
 
     Examples
     --------
@@ -560,7 +562,7 @@ def _regionprops(I, props, connectivity=2):
     a wrapper for skimage's regionprops. Not all of skimage's available blob
     properties are available here, but they can easily be added.
     Taken from RivGraph.im_utils
-    
+
     Note that 'perimeter' will only return the outer-most perimeter in the
     case of a region that contains holes.
 
@@ -630,37 +632,39 @@ def _regionprops(I, props, connectivity=2):
             for blob in coords:
                 # Crop to blob to reduce cv2 computation time and save memory
                 Ip, cropped = crop_binary_coords(blob)
-            
+
                 # Pad cropped image to avoid edge effects
                 Ip = np.pad(Ip, 1, mode="constant")
-                
+
                 # Get the perimeter using contours
-                contours_init = measure.find_contours(Ip, fully_connected='high',level=.99)
-                
+                contours_init = measure.find_contours(
+                    Ip, fully_connected="high", level=0.99
+                )
+
                 # In the cases of holes within the blob, multiple contours
                 # will be returned. We take the longest.
                 ci_lens = [len(ci) for ci in contours_init]
                 contours_init = contours_init[ci_lens.index(max(ci_lens))]
[…]
