### 2022-05-23 527e796
add coords_to_merit_tile tests
