### 2022-05-23 757f7e0
defensive programming for coords_to_merit_tile, #131
