### 2022-05-23 VeinsOfTheEarth/rabpro#130 @kanishkan91
> The capture command is a "magic command" that hides progress printing. https://ipython.readthedocs.io/en/stable/interactive/magics.html#cellmagic-capture

@jsta Thanks! Did not know about that.
