### 2022-05-23 3693d5b
add coords_to_merit_tile to basic ex, closes #131
