+++ _sources/examples/notebooks/basic_example.ipynb.txt
@@ -99,7 +99,7 @@
     },
     {
      "data": {
-      "image/png": "iVBORw0KGgoAAAANSUhEUgAAAPEAAAD4CAYAAAA9xADBAAAAOXRFWHRTb2Z0d2FyZQBNYXRwbG90bGliIHZlcnNpb24zLjUuMSwgaHR0cHM6Ly9tYXRwbG90bGliLm9yZy/YYfK9AAAACXBIWXMAAAsTAAALEwEAmpwYAAAnRElEQVR4nO3deXxU1dnA8d8zM8lkTyAJCYQlhH1VIAjIogKCgmLxfa1LtWpbl9ZSbau2dnOp2tb6Wmvt5tJarWtdqhUFVyiIgGHf9wAJWxII2ch+3j9mwCyTZJY7c+fOnO/nw8fMzL3nPol5cs6999zziFIKTdOsy2Z2AJqmBUYnsaZZnE5iTbM4ncSaZnE6iTXN4hxmB+BJRkaGys3NNTsMTQsba9asKVVKZXr6LCyTODc3l4KCArPD0LSwISL7O/pMD6c1zeJ0Emuaxekk1jSL00msaRank1jTLE4nsaZZnE5iTbM4ncSaZnE6iTXNRws3Hua6Z1fx28XbzQ4F0EmsaT55e30xv3h7M8t2lfLiqgPUNTaZHZJOYs18ZVV11DaYnwwdUUqhlOKDLUe4/ZX1lFXXA1Be08C6A+XmBkeYzp3WoseJ6nrOf3QJlbWNjMxJ4d/fmYzDHj59S3H5Kd5eX8yxijqeW1HY7vOa+sbQB9WGTmLNVH/4ZDeVta5E2FxcwSV/WE5KfAxOh41Yu40r8vtw0chsU2J78pNdvLWumOLyUzQ0eV6LrrahOcRRtaeTWDNNYWk1L6wsbPXe9iOVrV4P75ViShIv31XKcysKKa2q73S7cDgN0EmsmUIpxW8Wbe+whzut+MQpyqrq6J4Yi4iEJK6Pth3jpue9exRW98RaVKqobWDuE8s4ePxUl9u+ua6YN9cVE+uw0TM1jtG903jsq2cRE4Tz5mW7Svjhaxs4Vlnn9T7hcHVaJ7EWcve8scmrBG6pvrGZ/WU17C+rYW9JFeP6dePcARmGDLU3HCznlS8O8N6mI5w81eDTvh31xPtKq8lNTwjJ6CF8LgNqUaOwrDqg/bccquD5z/dz6z/X8On2YwG1tanoJN96voCXVx/0OYEBymvanzM3NjVzxyvrfOrRA6GTWAu5JKdxA8CFmw77ve/KvWXM++NySgJItudWFLK5+OSZ183NiqeX7WND0cmQ3UP26qcpIoVAJdAENCql8t3vLwC+CzQCC5VSd7fZbwjwaou38oBfKKUeDzhyzbJunpbHqn3HDWnreHXnV4/bKqms45YXCiguP8XRisB7yrrGZr71jwIGZSVRdOIUxSdOUd/kGmL//bN9Ibmy7sufxAuUUqWnX4jIBcBlwGilVJ2I9Gi7g1JqB3C2e3s7UAy8FVDEmuXNGJbFQ/NH8uC72zgV4C2asirvE7G5WbHg5bWsNbiHPFJRy5GK2nbvr9p3nDX7TzCuXzdDj9dWIMPpbwO/VkrVASilujo5mQHsUUp1uGqfFj2+NqEfC
[…]
