### 2022-05-24 VeinsOfTheEarth/rabpro#130 @jonschwenk
Contribution instructions are now in the main readme.

Re: Moving installation instructions main readme (second check box). We discussed this and decided that we prefer the current design as it is cleaner and prevents us from updating docs in two places.

`rabpro.utils.coords_to_merit_tile()` is now included in the Basic example.
