### 2022-05-24 b0f5c8d
add citations to readme
