### 2022-05-25 #4237 @kanishkan91
@jonschwenk I went through the latest draft of the paper after all edits. I guess one small edit I would recommend is that this package is useful for ESMs since it may help with initialization of basin level data. It might invite even more usage of this package.
