### 2022-05-25 270a461
print callstring on error
