### 2022-05-25 bfcafe7
debug vrt success check
