+++ rabpro/utils.py
@@ -379,7 +379,7 @@ def build_vrt(
     stdout, stderr = proc.communicate()
 
     # Check that vrt built successfully
-    if len(stderr) > 3:
+    if len(stderr) > 2:
         print(callstring)
         raise RuntimeError(f"Virtual raster did not build sucessfully. Error: {stderr}")
     else:
