### 2022-05-25 6936c9a
debug vrt success check
