### 2022-05-25 3dbf1e4
debug vrt success check
