### 2022-05-26 #4237 @crvernon
@jonschwenk we are almost there!

I found one thing that needs to be addressed in your paper:

- [ ] Line 101:  The link for "Full Example" is broken

- We want to make sure the archival has the correct metadata that JOSS requires.  This includes a title that matches the paper title and a correct author list.

So here is what we have left to do:
- [ ] Conduct a GitHub release of the current reviewed version of the software after the correction has been made above that you now have on the `main` and archive the reviewed software in Zenodo or a similar service (e.g., figshare, an institutional repository)
- [ ] Please list the DOI of the archived version here
