### 2022-05-26 #4237 @jsta
@jonschwenk We also need to make sure the code in the example notebooks reflect the current API
