+++ paper.bib
@@ -1,3 +1,24 @@
+@article{fisher2019parametric,
+  title={Parametric controls on vegetation responses to biogeochemical forcing in the CLM5},
+  author={Fisher, Rosie A and Wieder, William R and Sanderson, Benjamin M and Koven, Charles D and Oleson, Keith W and Xu, Chonggang and Fisher, Joshua B and Shi, Mingjie and Walker, Anthony P and Lawrence, David M},
+  journal={Journal of Advances in Modeling Earth Systems},
+  volume={11},
+  number={9},
+  pages={2879--2895},
+  year={2019},
+  publisher={Wiley Online Library}
+}
+
+@article{chen2020global,
+  title={Global land use for 2015--2100 at 0.05 resolution under diverse socioeconomic and climate scenarios},
+  author={Chen, Min and Vernon, Chris R and Graham, Neal T and Hejazi, Mohamad and Huang, Maoyi and Cheng, Yanyan and Calvin, Katherine},
+  journal={Scientific Data},
+  volume={7},
+  number={1},
+  pages={1--11},
+  year={2020},
+  publisher={Nature Publishing Group}
+}
 
 @article{gorelick_google_2017,
 	title = {Google {Earth} {Engine}: {Planetary}-scale geospatial analysis for everyone},
