### 2022-05-31 #4237 @crvernon
@jonschwenk and @jsta ... the missing DOIs above were not there in my last pass.  Could you see to these and generate a new release and post the version and archive here once resolved?
