+++ paper.bib
@@ -6,7 +6,8 @@
   number    = {9},
   pages     = {2879--2895},
   year      = {2019},
-  publisher = {Wiley Online Library}
+  publisher = {Wiley Online Library},
+  doi       = {10.1029/2019MS001609}
 }
 
 @article{chen2020global,
@@ -17,7 +18,8 @@
   number    = {1},
   pages     = {1--11},
   year      = {2020},
-  publisher = {Nature Publishing Group}
+  publisher = {Nature Publishing Group},
+  doi       = {10.1038/s41597-020-00669-x}
 }
 
 @article{gorelick_google_2017,
