### 2022-05-31 #4237 @crvernon
@jsta and @jonschwenk I still see an error in the new reference in the paper.  See line 56 on the PDF and please check for others that were recently added to make sure they compile correctly.  Once completed, let me know and I'll look again.  We will need a new version reflecting the changes.
